[refactor] orderer 컴포넌트 분리
